Refactor and clean up popup.js code
- Removed unused functions and consolidated event listeners for translation prompts.
- Simplified DOM interactions and initialization logic for Korean and English sessions.
